moving process_chunks
