+++ primitives/Cargo.toml
@@ -13,7 +13,6 @@ ark-serialize = "0.5.0"
 ark-std = { version = "0.5.0", features = ["parallel"] }
 ark-poly = { version = "0.5.0", features = ["parallel"] }
 sha2 = "0.10.8"
-crossbeam-channel = "0.5"
 num-traits = "0.2"
 thiserror = "2.0.11"
